requirements.txt
Flask==2.3.2
load-dotenv==0.1.0
# Flask-pyoidc==3.13.0
gevent==22.10.2
# greenlet==2.0.2
gunicorn==20.1.0
pytest==7.3.2
requests==2.32.3
coverage==7.5.4
